flask/requirements.txt
flask
flask-cors
np_utils
matplotlib
seaborn
pandas
graphviz
xgboost
catboost
vecstack
numpy

plotly
scikit-learn
lightgbm
openpyxl
joblib

#“unstructured[md,pdf,pptx]”
unstructured-client

# For MongoDB and FastAPI
pymongo
fastapi
uvicorn[standard]
unstructured
unstructured-inference
unstructured.pytesseract
urllib3
python-dotenv
panel
ipython
python-pptx
pdf2image
pdfminer
opencv-python

pikepdf
pypdf
psycopg2-binary
sqlalchemy
pyyaml
dash-ag-grid
pycountry
cartopy

# Deep Learning dependencies for image generation
torch
torchvision
torchaudio
Pillow
tqdm
diffusers
peft

streamlit